+++ src/components/film-details/film-details.jsx
@@ -2,6 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import Tabs from "../tabs/tabs.jsx";
 import FilmsList from "../films-list/films-list.jsx";
+import {MORE_LIKE_THIS_CARDS_COUNT} from "../../utils.js";
 
 const FilmDetails = (props) => {
   const {film, onCardAction, onImageAndTitleClick, filmsList} = props;
@@ -77,7 +78,7 @@ const FilmDetails = (props) => {
           <h2 className="catalog__title">More like this</h2>
           <FilmsList
             genre={film.genre}
-            filmsList={filteredFilms}
+            filmsList={filteredFilms.slice(0, MORE_LIKE_THIS_CARDS_COUNT)}
             onCardAction={onCardAction}
             onImageAndTitleClick={onImageAndTitleClick}
           />
