+++ src/utils.js
@@ -3,3 +3,5 @@ export const extend = (a, b) => {
 };
 
 export const DEFAULT_CARDS_COUNT = 8;
+export const INCREASER_CARDS_COUNT = 8;
+export const MORE_LIKE_THIS_CARDS_COUNT = 4;
